Add option BED file to SvPileup (#20)
* Add option BED file to SvPileup

The user may provide an optional BED file with `--targets-bed`.  The output breakpoints will be annotated with a comma-delimited list of targets in the BED file (by name) that overlap each side respectively.  Therefore, this changes the output format.

Furthermore, the `--targets-bed-requirement` option can be used to specify that only breakpoints that at least one target (at least one side via `OverlapsAny`, and both sides via `OverlapsBoth`) are written to the output.  In the case of `OverlapsBoth`, each side of a breakpoint does not have to overlap _the same_ target, just that they each independently overlap a target (can be a different target).